+++ CHANGES.rst
@@ -7,6 +7,8 @@ CHANGES
 * Don't suppress nested exception on timeout. Exception context points
   on cancelled line with suspended `await` (#13)
 
+* Introduce `.timeout` property (#16)
+
 1.2.1 (2017-05-02)
 ------------------